Rename type-translator.ts to type_translator.ts.
Style guide requires under_bar_case, which the rest of the project uses.
